==> Step3_DataCleaning.ipynb <==
import pandas as pd

df = pd.read_csv('/Users/pingyuanwang/Desktop/dat.csv')

df['X'] = t

t = df['C'] - df['O']

def f(row):
    if row['X'] > 0:
        val = 1
    else:
        val = 0
    return val

df['Y'] = df.apply(f, axis=1)

df2 = pd.DataFrame()
df2['Y'] = df['Y'][3:].tolist()
df2['X1'] = df['Y'][2:-1].tolist()
df2['X2'] = df['Y'][1:-2].tolist()
df2['X3'] = df['Y'][0:-3].tolist()

df2.to_csv('test_M1.csv',index = None, header = None)
